backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
SQLAlchemy==2.0.21
psycopg2-binary==2.9.9
redis==5.0.3
fpdf==1.7.2
pytest==8.2.1
python-dotenv==1.0.1
flask_sqlalchemy